Convert to experimental lua nvim setup
Nvim has recently introduced a way to configure it (and the ability to
create plugins) using lua. This updates my entire configuration to use
the lua style and removes coc-nvim entirely in favor of the new native
lsp engine in nvim and some autocomplete plugins.

It also makes a lot of updates so that all of the config files are
symlinked rather than copied. This makes it much easier to keep the
additions I make in sync.